Adopt coroutines/StateFlow, SavedStateHandle, and BigDecimal money
- Data layer: suspend getMenu() with withContext dispatcher

- ViewModel exposes uiState as StateFlow; menu loaded in viewModelScope

- Persist cart in SavedStateHandle so it survives process death

- Replace OrderViewModelFactory with viewModelFactory + createSavedStateHandle

- Fragments collect StateFlow/events via repeatOnLifecycle

- Model money as BigDecimal to avoid float rounding errors

+++ app/src/main/java/com/example/foodorder/data/model/MenuItem.kt
@@ -1,11 +1,16 @@
 package com.example.foodorder.data.model
 
+import java.math.BigDecimal
+
 /**
  * Represents a single item on the food menu.
+ *
+ * Money is modelled with [BigDecimal] rather than [Double] to avoid binary
+ * floating-point rounding errors in price calculations.
  */
 data class MenuItem(
     val id: Int,
     val name: String,
     val description: String,
-    val price: Double
+    val price: BigDecimal
 )
